// src/store/cylinderSize/cylinderSize.enum.ts
enum CylinderSizeTypeEnum {
  GET_CYLINDER_SIZE_SUCCESS = "GET_CYLINDER_SIZE_SUCCESS",
  GET_CYLINDER_SIZE_PENDING = "GET_CYLINDER_SIZE_PENDING",
  GET_CYLINDER_SIZE_FAILED = "GET_CYLINDER_SIZE_FAILED",

  UPDATE_CYLINDER_SIZE_SUCCESS = "UPDATE_CYLINDER_SIZE_SUCCESS",
  UPDATE_CYLINDER_SIZE_PENDING = "UPDATE_CYLINDER_SIZE_PENDING",
  UPDATE_CYLINDER_SIZE_FAILED = "UPDATE_CYLINDER_SIZE_FAILED",

  DELETE_CYLINDER_SIZE_SUCCESS = "DELETE_CYLINDER_SIZE_SUCCESS",
  DELETE_CYLINDER_SIZE_PENDING = "DELETE_CYLINDER_SIZE_PENDING",
  DELETE_CYLINDER_SIZE_FAILED = "DELETE_CYLINDER_SIZE_FAILED",

  ADD_CYLINDER_SIZE_SUCCESS = "ADD_CYLINDER_SIZE_SUCCESS",
  ADD_CYLINDER_SIZE_PENDING = "ADD_CYLINDER_SIZE_PENDING",
  ADD_CYLINDER_SIZE_FAILED = "ADD_CYLINDER_SIZE_FAILED",
}
export default CylinderSizeTypeEnum;